code refactoring and add auth - start
