+++ routes/api/auth.js
@@ -1,0 +1,6 @@
+const express = require("express");
+const router = express.Router();
+
+router.post("/register");
+
+module.exports = router;
