+++ server.js
@@ -10,13 +10,15 @@ const app = express();
 const {DB_HOST, PORT = 3000} = process.env;
 
 const formatsLogger = app.get("env") === "development" ? "dev" : "short";
-const contactsRouter = require("./api");
+const authRouter = require("./routes/api/auth");
+const contactsRouter = require("./routes/api/contacts");
 
 app.use(logger(formatsLogger));
 app.use(cors());
 app.use(express.json());
 
-app.use("/api", contactsRouter);
+app.use("/api/auth", authRouter);
+app.use("/api/contacts", contactsRouter);
 
 app.use((_, res, __) => {
   res.status(404).json({
